recovery: convert congestion control API to raw vtable
Using a trait for the congestion control operations means that different
algorithms will need to duplicate some of the logic due to not being
able to share common values between each other (e.g. congestion_window,
bytes_in_flight, ssthresh, ...). This also makes it hard to implement
more complex features (e.g. Hystart) that should really apply to all
algorithms.

This changes the congestion control API from a trait, where each
algorithm is implemented as a separate type that implements the common
trait, to a raw function vtabl, where each function is passed the entire
Recovery struct (so each algorithm don't have to define a separate type).

This allows all of the fields of Recovery to be available within each
congestion control implementation, so that they can be reused without
having to duplicate them, and without having to define awkward and
verbose trait methods to access them (another good side-effect being
that we don't need to pass many of these fields to the CC methods
anymore, as they can now be accessed directly).

This also makes it possible to mix and match the algorithm's methods
(e.g. in the future an implementation could reuse one of the existing
algorithms' functions), as well as wrap them and expand their logic
(commong logic between all algorithms can be implemented as part of
Recovery's on_packet_sent_cc(), on_packet_acked_cc() and
congestion_event() methods, rather than each algorithm's methods).

Due to the fact that Recovery's fields are private, the congestion
control implementations need to be submodule of recovery, rather than
having a separate cc submodule. One possibility would be to make a cc
module that is a submodule of recovery, but that doesn't seem to improve
things much, as the CC API is pretty small.